refactor: restructure project and redefine container styles
- Refactored project structure for better organization and maintainability
  - Moved `Header`, `Footer`, and `Layout` components into dedicated directories
  - Updated file structure for `SideBar` and `Wrapper` components
- Redefined container styles in `Wrappers.tsx` to improve layout consistency
- Minor adjustments to `user.tsx` and related components for alignment with new structure
