+++ components/layout/Layout/styled.tsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const LayoutWrapper = styled.div`
+  flex-grow: 1;
+  background-color: ${({ theme }) => theme.colors.primaryBg};
+`;
