+++ components/pages/user/Wrapper/index.tsx
@@ -1,0 +1,12 @@
+import { ReactNode } from "react";
+import { Container } from "./styled";
+
+interface WrapperProps {
+  children: ReactNode;
+}
+
+const Wrapper = ({ children }: WrapperProps) => {
+  return <Container>{children}</Container>;
+};
+
+export default Wrapper;
