+++ components/pages/user/Wrapper/styled.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+import { Container1344 } from "@/styles/container";
+
+export const Container = styled(Container1344)`
+  column-gap: 72px;
+  display: flex;
+`;
