+++ pages/user.tsx
@@ -1,9 +1,11 @@
 import React, { useState } from "react";
 import Profile from "@/components/pages/user/Profile";
 import SideBar from "@/components/pages/user/SideBar";
-import { UserContainer } from "@/styles/wrappers";
+// import { MainWrapper } from "@/styles/wrappers";
+import UserWrapper from "@/components/pages/user/Wrapper/index";
 import Orders from "@/components/pages/user/Orders";
 import Inquiries from "@/components/pages/user/Inquiries";
+import { MainWrapper } from "@/styles/wrappers";
 
 const User = () => {
   const [activeTab, setActiveTab] = useState<
@@ -12,12 +14,14 @@ const User = () => {
 
   return (
     <>
-      <UserContainer>
-        <SideBar setActiveTab={setActiveTab} activeTab={activeTab} />
-        {activeTab === "profile" && <Profile />}
-        {activeTab === "orders" && <Orders />}
-        {activeTab === "inquiries" && <Inquiries />}
-      </UserContainer>
+      <MainWrapper>
+        <UserWrapper>
+          <SideBar setActiveTab={setActiveTab} activeTab={activeTab} />
+          {activeTab === "profile" && <Profile />}
+          {activeTab === "orders" && <Orders />}
+          {activeTab === "inquiries" && <Inquiries />}
+        </UserWrapper>
+      </MainWrapper>
     </>
   );
 };
