+++ styles/Wrappers.tsx
@@ -1,9 +1,5 @@
 import styled from "styled-components";
 
-export const UserContainer = styled.div`
-  max-width: 1344px;
-  margin: 0px auto;
+export const MainWrapper = styled.div`
   padding: 80px 0;
-  column-gap: 72px;
-  display: flex;
 `;
